Fix typos.
Commit messages should in present tense (Fix rather than Fixed) and be searchable, that is, mention all class and variable names etc involved. It does not need to explain code (you should do that in the code comments), not intentions, but references to related external stuff (ie issue numbers, or discussions etc) is important.